refactor: folder and import structure
